fix pulse compression bugs, move CalbrateBase to a separate script, small changes re range_meter in CalibrateAZFP
